driver: dma: split header file differently
This patches just changes the header file organisation
